Wrote and called functions to return lists of all links in a buffer
created functions skrf-link-positions-in-buffer and
skrf-links-in-buffer.

rewrote rename-this-node-throughout-skrode to use
skrf-links-in-buffer, and break-skrode-link to use
skrf-link-positions-in-buffer.

changed handling of orphan nodes (nodes without working links) to
collect links to them in a special "skrode orphans" node.

fixed an args-out-of-range error from skrf-preview by replacing all instances
of save-excursion with save-mark-and-excursion.

and other minor code cleanup and fixes.