replaced all calls to skrf-link-to-text with skrf-remove-link-delims
which is a simpler function that deletes all link delims anywhere in a
string, and does not require there to be any.

deleted no longer used function skrf-link-to-text.

changed some uses of skrode-node-name that could potentially be
called in eg a temp buffer without buffer-local variables to use the
function skrf-node-name.