fixed so skrode mode should work with annotations eg longlines.el!
by making sure break-skrode-link does nothing if it was called with a buffer not in skrode mode.

also: rewrote with-inhibit-modification-hooks to be shorter & simpler.

moved skrf-new-window back out of skrf-open-node-in-new-window, because it is occasionally also called by skrf-open-node-in-same-window.

created yank handlers to strip properties from skrode links and skrode name when copy-pasting.

added skrode-name text property to the skrode name for yank handler to handle it by.